requirements.txt
python-dotenv==1.0.0
tweepy==4.14.0
google-generativeai==0.3.1
markdown==3.5.1
PyYAML==6.0.1
